Move localVue creation to utils.ts
