+++ present_ideas/present_ideas/static/pages/__tests__/LandingPage.test.js
@@ -1,17 +1,13 @@
 import { render } from "@testing-library/vue";
-import { createLocalVue } from "@vue/test-utils";
-import Vuex from "vuex";
 import getStore from "../../store";
 import LandingPage from "../LandingPage.vue";
-
-const localVue = createLocalVue();
-localVue.use(Vuex);
+import { getVueWithVuex } from "./utils";
 
 describe("LandingPage", () => {
   it("matches snapshot", () => {
-    const store = getStore();
     expect(
-      render(LandingPage, { localVue, store }).container
+      render(LandingPage, { localVue: getVueWithVuex(), store: getStore() })
+        .container
     ).toMatchSnapshot();
   });
 });
