--- present_ideas/present_ideas/static/pages/__tests__/mockRoute.ts
@@ -1,1 +1,0 @@
-export const mockRoute = { params: {} };
