+++ present_ideas/present_ideas/static/pages/__tests__/utils.ts
@@ -1,0 +1,9 @@
+import { createLocalVue } from "@vue/test-utils";
+import Vuex from "vuex";
+
+export const mockRoute = { params: {} };
+export const getVueWithVuex = () => {
+  const localVue = createLocalVue();
+  localVue.use(Vuex);
+  return localVue;
+};
